fix(admin-hooks): split FetchAuthedError into client-safe module
useAdminMutation is a client hook but FetchAuthedError lived in
lib/fetch/authed.ts which imports next/headers (server-only).
Extract the error classes to lib/fetch/errors.ts and re-export from
authed.ts to keep the existing server-side import surface.
