+++ apps/web/src/lib/fetch/authed.ts
@@ -4,6 +4,10 @@
 
 import { cookies } from "next/headers";
 
+import { AuthRequiredError, FetchAuthedError } from "./errors";
+
+export { AuthRequiredError, FetchAuthedError };
+
 const FALLBACK_INTERNAL_API = "http://127.0.0.1:8787";
 
 const resolveApiBase = (): string => {
@@ -13,24 +17,6 @@ const resolveApiBase = (): string => {
   if (pub && pub.length > 0) return pub.replace(/\/$/, "");
   return FALLBACK_INTERNAL_API;
 };
-
-export class AuthRequiredError extends Error {
-  constructor(message = "AUTH_REQUIRED") {
-    super(message);
-    this.name = "AuthRequiredError";
-  }
-}
-
-export class FetchAuthedError extends Error {
-  readonly status: number;
-  readonly bodyText: string;
-  constructor(status: number, bodyText: string) {
-    super(`fetchAuthed failed: ${status}`);
-    this.name = "FetchAuthedError";
-    this.status = status;
-    this.bodyText = bodyText;
-  }
-}
 
 const buildCookieHeader = async (): Promise<string> => {
   const store = await cookies();
